+++ app/src/main/kotlin/com/d4rk/englishwithlidia/plus/ui/lessons/LessonsFragment.kt
@@ -9,12 +9,15 @@ import androidx.lifecycle.ViewModelProvider
 import com.d4rk.englishwithlidia.plus.R
 import com.d4rk.englishwithlidia.plus.databinding.FragmentLessonsBinding
 import com.d4rk.englishwithlidia.plus.ui.viewmodel.ViewModel
+import me.zhanghai.android.fastscroll.FastScrollerBuilder
+
 class LessonsFragment : Fragment(R.layout.fragment_lessons) {
     private var _binding: FragmentLessonsBinding? = null
     private val binding get() = _binding!!
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         ViewModelProvider(this)[ViewModel::class.java]
         _binding = FragmentLessonsBinding.inflate(inflater, container, false)
+        FastScrollerBuilder(binding.lessonsScrollView).useMd2Style().build()
         binding.lessonsLesson1Card.setOnClickListener {
             val intent = Intent(activity, Lesson1Activity::class.java)
             startActivity(intent)