+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.4_r1:
+- Added language preferences (Android 13+);
+- Fixed Firebase crashes;
+- Fixed splash screen;
+- Removed language section from settings;
+- Removed unused resources;
+- New scroll bars;
+- Bug fixes;
+- Tweaks;
+- Under the hood improvements.
+
 # Version 1.3_r1:
 - Target SDK now is 26;
 - Removed unused resources;